Factor out migrator
